requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1       # ASGI server; [standard] adds websockets + httptools

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36              # ORM + query builder
# Uncomment if you switch to PostgreSQL:
# psycopg2-binary==2.9.10

# ── Authentication ────────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4          # Password hashing
python-jose[cryptography]==3.3.0  # JWT encode / decode
bcrypt==4.2.1

# ── File uploads ──────────────────────────────────────────────────────────────
python-multipart==0.0.18        # Required for FastAPI UploadFile
PyMuPDF==1.25.1                 # PDF text extraction (fitz)
python-docx==1.1.2              # DOCX text extraction

# ── Claude API ────────────────────────────────────────────────────────────────
anthropic==0.42.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1            # Load .env in development
pydantic[email]==2.10.3         # Data validation (email validator extra)
